requirements.txt
fastapi==0.109.0
uvicorn==0.24.0
requests==2.31.0
transformers==4.37.2
annoy==1.17.3
tqdm==4.66.2
pydantic==2.6.1
gensim==4.3.2
scipy==1.10.1
fasttext-wheel==0.9.2
nltk==3.8.1
huggingface-hub==0.20.3
pymorphy2==0.9.1
elasticsearch==8.11.1
sentence-transformers==2.3.1
torch==2.1.2
openai==1.63.2